Refactors diona split code for maximum reuse.

+++ code/modules/mob/living/carbon/human/human_powers.dm
@@ -224,7 +224,9 @@
 	set name = "Split"
 	set desc = "Split your humanoid form into its constituent nymphs."
 	set category = "Abilities"
+	diona_split_into_nymphs(5, 1)	// Separate proc to void argments being supplied when used as a verb
 
+/mob/living/carbon/human/proc/diona_split_into_nymphs(var/number_of_resulting_nymphs, var/delete_self = 1)
 	var/turf/T = get_turf(src)
 
 	var/mob/living/carbon/alien/diona/S = new(T)
@@ -243,8 +245,8 @@
 		D.loc = T
 		D.set_dir(pick(NORTH, SOUTH, EAST, WEST))
 
-	if(nymphs < 5)
-		for(var/i in nymphs to 4)
+	if(nymphs < number_of_resulting_nymphs)
+		for(var/i in nymphs to (number_of_resulting_nymphs - 1))
 			var/mob/M = new /mob/living/carbon/alien/diona(T)
 			M.set_dir(pick(NORTH, SOUTH, EAST, WEST))
 
@@ -254,4 +256,5 @@
 
 	visible_message("<span class='warning'>\The [src] quivers slightly, then splits apart with a wet slithering noise.</span>")
 
-	qdel(src)
+	if(delete_self)
+		qdel(src)
